ref: extract common functions getMyInfoPrefix and getAnswersForChild to shared submission.utils
